avance

+++ db/migrate/19250203212244_create_sectors.rb
@@ -2,7 +2,7 @@ class CreateSectors < ActiveRecord::Migration[7.1]
   def change
     create_table :sectors do |t|
       t.string :nombre
-      t.integer :hectareas
+      t.decimal :hectareas
 
       t.timestamps
     end
